fix: 로직 경로 수정
